refactor: separated app routes and business logic
